+++ src/input/combobox/utils.js
@@ -1,3 +1,5 @@
+import { alignItem } from '../../utils';
+
 export function groupData (items) {
 	const nogroup = [];
 	const _groups = {};
@@ -12,7 +14,7 @@ export function groupData (items) {
 }
 
 
-export function highlight (listEl) {
+export function scrollToSelectedItem (listEl) {
 	if (!listEl) return;
 	requestAnimationFrame(() => {
 		const selectedEl = listEl.querySelector('.selected');
@@ -69,4 +71,42 @@ export function findValueInSource (val, items) {
 	if (!val) return val;
 	if (!Array.isArray(val)) return findSourceItem(val, items);
 	return val.map(v => findSourceItem(v, items));
+}
+
+
+export function getInputValue (_val, isMultiselect = false) {
+	if (!isMultiselect) return _val?.name || _val || '';
+	if (!Array.isArray(_val)) _val = [_val];
+	return _val.map(i => i.name || i).join(', ');
+}
+
+
+export function alignDropdown (listElement, inputElement, e) {
+	requestAnimationFrame(() => {
+		alignItem({
+			element: listElement,
+			target: inputElement,
+			setMinWidthToTarget: true,
+			offsetH: -1
+		});
+		if (e && e.type === 'focus') inputElement.select();
+	});
+}
+
+
+function hasSingleValueChanged (oldV, newV) {
+	return (oldV.id || oldV.name || oldV) !== (newV.id || newV.name || newV);
+}
+
+
+export function hasValueChanged (oldV, newV, multiselect = false) {
+	if (!multiselect) return hasSingleValueChanged(oldV, newV);
+
+	if (!Array.isArray(oldV)) oldV = [oldV];
+	if (!Array.isArray(newV)) newV = [newV];
+	if (oldV.length !== newV.length) return true;
+	for (let i = 0; i < newV.length; i++) {
+		if (hasSingleValueChanged(oldV[i], newV[i])) return true;
+	}
+	return false;
 }