+++ tests/Table.spec.js
@@ -1,7 +1,7 @@
 import { render, fireEvent } from '@testing-library/svelte';
+import { default as userEvent } from '@testing-library/user-event';
 import jest from 'jest-mock';
 import Table from './helpers/Table.svelte';
-import userEvent from '@testing-library/user-event';
 
 
 test('Table', async () => {