+++ tests/input/InputDate.spec.js
@@ -1,5 +1,5 @@
 import { render, fireEvent } from '@testing-library/svelte';
-import userEvent from '@testing-library/user-event';
+import { default as userEvent } from '@testing-library/user-event';
 import jest from 'jest-mock';
 
 import { InputDate } from '../../src/input/input-date';