+++ tests/input/InputPassword.spec.js
@@ -1,5 +1,5 @@
 import { fireEvent, render } from '@testing-library/svelte';
-import userEvent from '@testing-library/user-event';
+import { default as userEvent } from '@testing-library/user-event';
 import jest from 'jest-mock';
 import zxcvbn from 'zxcvbn';